+++ modules/auth/match.go
@@ -4,9 +4,10 @@ import (
 	"context"
 
 	"github.com/spaceuptech/space-cloud/config"
-	"github.com/spaceuptech/space-cloud/crud"
 	"github.com/spaceuptech/space-cloud/model"
 	"github.com/spaceuptech/space-cloud/utils"
+
+	"github.com/spaceuptech/space-cloud/modules/crud"
 )
 
 func (m *Module) matchRule(rule *config.Rule, args map[string]interface{}) error {
