+++ modules/crud/mgo/delete.go
