+++ modules/filestore/local/delete.go
